Initial support for Ferry Boats
Treats "ferry-boat" entities the same as "boat" entities, but they can only go on waterway rails.

There is still something weird because the ferry-boat wagon flips sometimes when loading, but the boat-engine locomotive doesn't move.  The cargo-ships logic gets confused and doesn't delete the boat-engine when it's on the wrong end of the boat.